requirements.txt
biopython==1.79
matplotlib==3.9.2
openpyxl==3.0.9
pandas==1.4.1
scipy==1.14.1
snakemake==7.25.0
